Worked on Goal page Api Connection
I made some modifications to goal-page.tsx, package_lock.json, and serializer.py to solve issues in the API connection. Now, the "Set New Goal" button tries to send data to the backend, but the process is not terminated due to an unknown reason. I will continue working on this page and the API connection.